Multiple enemies spawn
The method spawns enemies in a given row and column number. Enemies are always aligned to the middle. I used the value 5, 11, which barely fits the screen, however, this is the original Space Invaders row and col so I'll let you decide if you will lower the image size or have fewer columns